final fixes to Flatty, release 1.0.7

+++ flatty/CompressionLevel.swift
@@ -11,16 +11,16 @@ import ZipArchive
 public enum CompressionLevel: Int {
     case `default` = 0, fast, normal, best
 
-    public var cValue: Int32 {
+    public var cValue: Int16 {
         switch self {
         case .`default`:
-            return MZ_COMPRESS_LEVEL_DEFAULT
+            return Int16(MZ_COMPRESS_LEVEL_DEFAULT)
         case .fast:
-            return MZ_COMPRESS_LEVEL_FAST
+            return Int16(MZ_COMPRESS_LEVEL_FAST)
         case .normal:
-            return MZ_COMPRESS_LEVEL_NORMAL
+            return Int16(MZ_COMPRESS_LEVEL_NORMAL)
         case .best:
-            return MZ_COMPRESS_LEVEL_BEST
+            return Int16(MZ_COMPRESS_LEVEL_BEST)
         }
     }
 }
